xDS interop: Move k8s-specific logic out of the test app
