+++ tools/run_tests/xds_k8s_test_driver/framework/xds_url_map_test_resources.py
@@ -28,6 +28,8 @@ from framework.infrastructure import k8s
 from framework.infrastructure import traffic_director
 from framework.test_app import client_app
 from framework.test_app import server_app
+import framework.test_app.runners.k8s.k8s_xds_client_runner
+import framework.test_app.runners.k8s.k8s_xds_server_runner
 
 flags.adopt_module_key_flags(xds_flags)
 flags.adopt_module_key_flags(xds_k8s_flags)
@@ -165,7 +167,7 @@ class GcpResourceManager(metaclass=_MetaSingletonAndAbslFlags):
         self.k8s_namespace = k8s.KubernetesNamespace(self.k8s_api_manager,
                                                      self.resource_prefix)
         # Kubernetes Test Servers
-        self.test_server_runner = server_app.KubernetesServerRunner(
+        self.test_server_runner = framework.test_app.runners.k8s.k8s_xds_server_runner.KubernetesServerRunner(
             self.k8s_namespace,
             deployment_name=self.server_name,
             image_name=self.server_image,
@@ -176,7 +178,7 @@ class GcpResourceManager(metaclass=_MetaSingletonAndAbslFlags):
             xds_server_uri=self.xds_server_uri,
             network=self.network,
             enable_workload_identity=self.enable_workload_identity)
-        self.test_server_alternative_runner = server_app.KubernetesServerRunner(
+        self.test_server_alternative_runner = framework.test_app.runners.k8s.k8s_xds_server_runner.KubernetesServerRunner(
             self.k8s_namespace,
             deployment_name=self.server_name + '-alternative',
             image_name=self.server_image,
@@ -188,7 +190,7 @@ class GcpResourceManager(metaclass=_MetaSingletonAndAbslFlags):
             network=self.network,
             enable_workload_identity=self.enable_workload_identity,
             reuse_namespace=True)
-        self.test_server_affinity_runner = server_app.KubernetesServerRunner(
+        self.test_server_affinity_runner = framework.test_app.runners.k8s.k8s_xds_server_runner.KubernetesServerRunner(
             self.k8s_namespace,
             deployment_name=self.server_name + '-affinity',
             image_name=self.server_image,
@@ -211,10 +213,11 @@ class GcpResourceManager(metaclass=_MetaSingletonAndAbslFlags):
         logging.info('GcpResourceManager: client_namespace_suffix=%s',
                      client_namespace_suffix)
         # Kubernetes Test Client
-        return client_app.KubernetesClientRunner(
+        return framework.test_app.runners.k8s.k8s_xds_client_runner.KubernetesClientRunner(
             k8s.KubernetesNamespace(
                 self.k8s_api_manager,
-                client_app.KubernetesClientRunner.make_namespace_name(
+                framework.test_app.runners.k8s.k8s_xds_client_runner.
+                KubernetesClientRunner.make_namespace_name(
                     self.resource_prefix, client_namespace_suffix)),
             deployment_name=self.client_name,
             image_name=self.client_image,
