+++ tools/run_tests/xds_k8s_test_driver/tests/remove_neg_test.py
@@ -20,6 +20,7 @@ from absl.testing import absltest
 from framework import xds_k8s_testcase
 from framework.infrastructure import k8s
 from framework.test_app import server_app
+import framework.test_app.runners.k8s.k8s_xds_server_runner
 
 logger = logging.getLogger(__name__)
 flags.adopt_module_key_flags(xds_k8s_testcase)
@@ -33,7 +34,7 @@ class RemoveNegTest(xds_k8s_testcase.RegularXdsKubernetesTestCase):
 
     def setUp(self):
         super().setUp()
-        self.alternate_server_runner = server_app.KubernetesServerRunner(
+        self.alternate_server_runner = framework.test_app.runners.k8s.k8s_xds_server_runner.KubernetesServerRunner(
             k8s.KubernetesNamespace(self.k8s_api_manager,
                                     self.server_namespace),
             deployment_name=self.server_name + '-alt',
